navigation
tried my hand at adding navigation but it's not quite there.

aiming to have a big map_list with all of the location.names in it so that we can use a similar system to the dialogue choice options. that way the command for movement is ALWAYS move # and if the # is invalid then it won't move the player.